Move API endpoint defaults to config.constants submodule
This is to eliminate config submodule dependency on api submodule.
