+++ dwave/cloud/config/__init__.py
@@ -158,5 +158,6 @@ Examples:
     >>> client.close() # doctest: +SKIP
 """
 
+from .constants import *
 from .loaders import *
 from .models import *
