+++ dwave/cloud/regions.py
@@ -18,7 +18,7 @@ from typing import Dict, List, Optional, Union
 from urllib.parse import urlsplit
 
 from dwave.cloud import api
-from dwave.cloud.api.constants import (
+from dwave.cloud.config.constants import (
     DEFAULT_REGION, DEFAULT_SOLVER_API_ENDPOINT, DEFAULT_LEAP_API_ENDPOINT,
     DEFAULT_METADATA_API_ENDPOINT)
 from dwave.cloud.config.models import ClientConfig, validate_config_v1
